+++ pkg/validation/app_test.go
@@ -101,7 +101,57 @@ func Test_ValidateApp(t *testing.T) {
 			},
 		},
 		{
-			name: "case 2: missing version label",
+			name: "case 2: flawless flow with `v` prefixed version",
+			obj: v1alpha1.App{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "kiam",
+					Namespace: "eggs2",
+					Labels: map[string]string{
+						label.AppOperatorVersion: "2.6.0",
+					},
+				},
+				Spec: v1alpha1.AppSpec{
+					Catalog:   "giantswarm",
+					Name:      "kiam",
+					Namespace: "kube-system",
+					Config: v1alpha1.AppSpecConfig{
+						ConfigMap: v1alpha1.AppSpecConfigConfigMap{
+							Name:      "eggs2-cluster-values",
+							Namespace: "eggs2",
+						},
+					},
+					KubeConfig: v1alpha1.AppSpecKubeConfig{
+						Context: v1alpha1.AppSpecKubeConfigContext{
+							Name: "eggs2-kubeconfig",
+						},
+						InCluster: false,
+						Secret: v1alpha1.AppSpecKubeConfigSecret{
+							Name:      "eggs2-kubeconfig",
+							Namespace: "eggs2",
+						},
+					},
+					UserConfig: v1alpha1.AppSpecUserConfig{
+						ConfigMap: v1alpha1.AppSpecUserConfigConfigMap{
+							Name:      "kiam-user-values",
+							Namespace: "eggs2",
+						},
+					},
+					Version: "v1.4.0",
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("giantswarm", "default"),
+			},
+			configMaps: []*corev1.ConfigMap{
+				newTestConfigMap("eggs2-cluster-values", "eggs2"),
+				newTestConfigMap("kiam-user-values", "eggs2"),
+			},
+			secrets: []*corev1.Secret{
+				newTestSecret("eggs2-kubeconfig", "eggs2"),
+			},
+		},
+		{
+			name: "case 3: missing version label",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -123,7 +173,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: label `app-operator.giantswarm.io/version` not found",
 		},
 		{
-			name: "case 3: spec.catalog not found",
+			name: "case 4: spec.catalog not found",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -148,7 +198,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: catalog `missing` not found",
 		},
 		{
-			name: "case 4: spec.config.configMap not found",
+			name: "case 5: spec.config.configMap not found",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -179,7 +229,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "app config map not found error: configmap `dex-app-values` in namespace `giantswarm` not found",
 		},
 		{
-			name: "case 5: spec.config.configMap no namespace specified",
+			name: "case 6: spec.config.configMap no namespace specified",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -210,7 +260,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: namespace is not specified for configmap `dex-app-values`",
 		},
 		{
-			name: "case 6: spec.config.secret not found",
+			name: "case 7: spec.config.secret not found",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -241,7 +291,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: secret `dex-app-secrets` in namespace `giantswarm` not found",
 		},
 		{
-			name: "case 7: spec.config.secret no namespace specified",
+			name: "case 8: spec.config.secret no namespace specified",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -272,7 +322,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: namespace is not specified for secret `dex-app-secrets`",
 		},
 		{
-			name: "case 8: spec.kubeConfig.secret not found",
+			name: "case 9: spec.kubeConfig.secret not found",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",
@@ -304,7 +354,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "kube config not found error: kubeconfig secret `eggs2-kubeconfig` in namespace `eggs2` not found",
 		},
 		{
-			name: "case 9: spec.kubeConfig.secret no namespace specified",
+			name: "case 10: spec.kubeConfig.secret no namespace specified",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",
@@ -336,7 +386,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: namespace is not specified for kubeconfig secret `eggs2-kubeconfig`",
 		},
 		{
-			name: "case 10: spec.userConfig.configMap not found",
+			name: "case 11: spec.userConfig.configMap not found",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -367,7 +417,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: configmap `dex-app-user-values` in namespace `giantswarm` not found",
 		},
 		{
-			name: "case 11: spec.userConfig.configMap no namespace specified",
+			name: "case 12: spec.userConfig.configMap no namespace specified",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -398,7 +448,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: namespace is not specified for configmap `dex-app-user-values`",
 		},
 		{
-			name: "case 12: spec.userConfig.secret not found",
+			name: "case 13: spec.userConfig.secret not found",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -429,7 +479,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: secret `dex-app-user-secrets` in namespace `giantswarm` not found",
 		},
 		{
-			name: "case 13: spec.userConfig.secret no namespace specified",
+			name: "case 14: spec.userConfig.secret no namespace specified",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "dex-app-unique",
@@ -460,7 +510,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: namespace is not specified for secret `dex-app-user-secrets`",
 		},
 		{
-			name: "case 14: spec.userConfig.configMap.name incorrect for default catalog app",
+			name: "case 15: spec.userConfig.configMap.name incorrect for default catalog app",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",
@@ -491,7 +541,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: user configmap must be named `kiam-user-values` for app in default catalog",
 		},
 		{
-			name: "case 15: spec.userConfig.secret.name incorrect for default catalog app",
+			name: "case 16: spec.userConfig.secret.name incorrect for default catalog app",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",
@@ -522,7 +572,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: user secret must be named `kiam-user-secrets` for app in default catalog",
 		},
 		{
-			name: "case 16: metadata.name exceeds max length",
+			name: "case 17: metadata.name exceeds max length",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "cluster-autoscaler-1.2.2-2b060b8bda545a7b6aeff1b8cb13951181ae30d3",
@@ -547,7 +597,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: name `cluster-autoscaler-1.2.2-2b060b8bda545a7b6aeff1b8cb13951181ae30d3` is 65 chars and exceeds max length of 53 chars",
 		},
 		{
-			name: "case 17: legacy version label is rejected",
+			name: "case 18: legacy version label is rejected",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",
@@ -572,7 +622,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: label `app-operator.giantswarm.io/version` has invalid value `1.0.0`",
 		},
 		{
-			name: "case 18: nginx user values configmap name is not restricted",
+			name: "case 19: nginx user values configmap name is not restricted",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",
@@ -605,7 +655,7 @@ func Test_ValidateApp(t *testing.T) {
 			},
 		},
 		{
-			name: "case 19: spec.kubeConfig.secret no name specified",
+			name: "case 20: spec.kubeConfig.secret no name specified",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",
@@ -637,7 +687,7 @@ func Test_ValidateApp(t *testing.T) {
 			expectedErr: "validation error: name is not specified for kubeconfig secret",
 		},
 		{
-			name: "case 20: skip validation when giantswarm.io/managed-by label equals flux",
+			name: "case 21: skip validation when giantswarm.io/managed-by label equals flux",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",
@@ -1020,7 +1070,46 @@ func Test_ValidateNamespace(t *testing.T) {
 			},
 		},
 		{
-			name: "case 1: namespace annotation collision",
+			name: "case 1: flawless with `v` version prefixed",
+			obj: v1alpha1.App{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "kiam",
+					Namespace: "eggs2",
+				},
+				Spec: v1alpha1.AppSpec{
+					Catalog:   "giantswarm",
+					Name:      "kiam",
+					Namespace: "kube-system",
+					NamespaceConfig: v1alpha1.AppSpecNamespaceConfig{
+						Annotations: map[string]string{
+							"linkerd.io/inject": "enabled",
+						},
+					},
+					Version: "v1.4.0",
+				},
+			},
+			apps: []*v1alpha1.App{
+				{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "another-kiam-1",
+						Namespace: "eggs2",
+					},
+					Spec: v1alpha1.AppSpec{
+						Catalog:   "giantswarm",
+						Name:      "kiam",
+						Namespace: "kube-system",
+						NamespaceConfig: v1alpha1.AppSpecNamespaceConfig{
+							Annotations: map[string]string{
+								"linkerd.io/inject": "enabled",
+							},
+						},
+						Version: "1.3.0-rc1",
+					},
+				},
+			},
+		},
+		{
+			name: "case 2: namespace annotation collision",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",
@@ -1060,7 +1149,7 @@ func Test_ValidateNamespace(t *testing.T) {
 			expectedErr: "app `kiam` annotation `linkerd.io/inject` for target namespace `kube-system` collides with value `disabled` for app `another-kiam-1`",
 		},
 		{
-			name: "case 2: namespace label collision",
+			name: "case 3: namespace label collision",
 			obj: v1alpha1.App{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      "kiam",