+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support for the `v` prefixed version. This enables Flux to use `v`-prefixed image tags as Apps version to automatically update them.
+- Support for App CRs with a `v` prefixed version. This enabled Flux to automatically update the version based on its image tag. 
 
 ## [5.6.0] - 2021-11-29