+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.6.1] - 2021-12-08
+
 ## [6.0.0] - 2021-11-29
 
 - Drop `apiextensions` dependency in favor of `apiextensions-application`.
@@ -275,7 +277,8 @@ since there can be a delay creating them on cluster creation.
 
 - First release
 
-[Unreleased]: https://github.com/giantswarm/app/compare/v6.0.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app/compare/v5.6.1...HEAD
+[5.6.1]: https://github.com/giantswarm/app/compare/v6.0.0...v5.6.1
 [6.0.0]: https://github.com/giantswarm/app/compare/v5.6.0...v6.0.0
 [5.6.0]: https://github.com/giantswarm/app/compare/v5.5.0...v5.6.0
 [5.5.0]: https://github.com/giantswarm/app/compare/v5.4.0...v5.5.0